Creation des classes de services GestionEntreprise et Gestion Ordinateur pour eviter que l'interface utilisateur est acces au classe DAO dirrectement.

+++ src/com/cdb/dao/InstanceOrdinateurDao.java
@@ -14,10 +14,10 @@ public interface InstanceOrdinateurDao {
 
 	List<Ordinateur> findOrdinateurByPage(int numeroPage, int ligneParPage) throws ConnexionDatabaseException, RequeteQueryException;
 
-	Ordinateur findOrdinateurByID(int index) throws ConnexionDatabaseException, RequeteQueryException;
+	Ordinateur findOrdinateurByID(long index) throws ConnexionDatabaseException, RequeteQueryException;
 
 	void updateOrdinateur(Ordinateur ordinateur) throws ConnexionDatabaseException, RequeteQueryException;
 
-	void suppressionOrdinateur(int index) throws ConnexionDatabaseException, RequeteQueryException;
+	void suppressionOrdinateur(long index) throws ConnexionDatabaseException, RequeteQueryException;
 
 }
